+++ src/Playbook.Core/Players/TrendDirection.cs
@@ -1,0 +1,8 @@
+namespace Playbook.Core.Players;
+
+public enum TrendDirection
+{
+    Up = 0,
+    Flat = 1,
+    Down = 2
+}